chore: Upgrade `@metamask/eth-query` to `v4.0.0`
The package @metamask/eth-query` has been upgraded to `v4.0.0`. The
breaking change in v4 was a change to the `Provider` type, to make it
narrower so that it matches the type we use in other packages. This is
not anticipated to require any changes.

Changelog: https://github.com/MetaMask/eth-query/blob/main/CHANGELOG.md#400