chore: Upgrade `@metamask/token-tracker` to `v7.0.1`
The package `@metamask/token-tracker` has been upgraded from v4 to
`v7.0.1. The breaking changes were mostly bumps to the minimum Node.js
version, which doesn't impact this project because we're still above
the current minimum in v7.

The only other breaking change was an update to the block tracker,
which is exposed as a public property of the token tracker class. We
don't use that property in the extension, so this change has no impact.

Changelog: https://github.com/MetaMask/eth-token-tracker/blob/main/CHANGELOG.md#701---2023-12-20